streamlit_app/requirements.txt
streamlit==1.33.0
requests==2.31.0
Pillow==10.3.0
matplotlib==3.8.4
mlflow==2.13.0